Remove Riak support
Removed the class JsonStoreConfig, as we don't have any other
Json config except Cloudant

BugzID: 56722
