+++ README.md
@@ -1,31 +1,20 @@
 Spark SQL Cloudant External Datasource
 ================
 
-Cloudant integration with Spark as Spark SQL external datasource. Most of the implementation can be re-used by other JsonStore, e.g. Riak. 
-
+Cloudant integration with Spark as Spark SQL external datasource. 
 
 ## Contents:
 
 ### RelationProvider implementations
 
-All implementions are under [here] (cloudant-spark-sql/src/main/scala/com/cloudant/spark)
+All implementions can be found [here] (cloudant-spark-sql/src/main/scala/com/cloudant/spark):
 
-#### Cloudant
-
-Relation Provider Name | Table Option | Scan Type | Column Pruning | Predicates Push Down | Parallel Loading | Insertable | Source Code Location
---- | --- | --- | --- | --- | --- | --- | --- 
-com.cloudant.spark.DefaultSource|database or path, index|PrunedFilteredScan| Yes |_id or first predicate | Yes, except with index | Yes | [DefaultSource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/DefaultSource.scala)
-com.cloudant.spark.CloudantRP|database|TableScan| No | No | No | No | [CloudantDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/CloudantDatasource.scala)
-com.cloudant.spark.CloudantPrunedFilteredRP|database|PrunedFilteredScan| Yes |_id or first predicate | No | No | [CloudantPrunedFilteredDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/CloudantPrunedFilteredDatasource.scala)
-com.cloudant.spark.CloudantPartitionedPrunedFilteredRP|database, index|PrunedFilteredScan| Yes |_id or first predicate |  Yes, except with index  | No |[CloudantPartitionedPrunedFilteredDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/CloudantPartitionedPrunedFilteredDatasource.scala)
-
-
-#### Riak
-
-Relation Provider Name | Table Option | Scan Type | Column Pruning | Predicates Push Down | Parallel Loading | Insertable | Source Code Location
---- | --- | --- | --- | --- | --- | --- | ---
-com.cloudant.spark.CloudantPartitionedPrunedFilteredRP|path|PrunedFilteredScan| Yes | first predicate | Yes | No |[RiakPartitionedPrunedFilteredDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/riak/RiakPartitionedPrunedFilteredDatasource.scala)
-
+Relation Provider Name | Table Option | Scan Type | Column Pruning | Predicates Push Down | Parallel Loading | Insertable | 
+--- | --- | --- | --- | --- | --- | --- | 
+[DefaultSource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/DefaultSource.scala)|database or path, index|PrunedFilteredScan| Yes |_id or first predicate | Yes, except with index | Yes |
+[CloudantDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/CloudantDatasource.scala)|database|TableScan| No | No | No | No | 
+[CloudantPrunedFilteredDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/)|database|PrunedFilteredScan| Yes |_id or first predicate | No | No | CloudantPrunedFilteredDatasource.scala)
+[CloudantPartitionedPrunedFilteredDatasource.scala](cloudant-spark-sql/src/main/scala/com/cloudant/spark/CloudantPartitionedPrunedFilteredDatasource.scala)|database, index|PrunedFilteredScan| Yes |_id or first predicate |  Yes, except with index  | No |
 
 
 ### Binary download:
@@ -151,11 +140,6 @@ Spark Version | Release # | Binary Location
 
 [Sample code on using DataFrame option to define cloudant configuration](spark-test/src/main/scala/mytest/spark/CloudantDFOption.scala)
 
-### Note: 
-	
-In the above table creation, you can replace "USING com.cloudant.spark.CloudantRP" with other RelationProvider implementations
-
-	
 	
 ## Job Submission
 
@@ -184,8 +168,6 @@ Name | Default | Meaning
 cloudant.host||cloudant host url
 cloudant.username||cloudant userid
 cloudant.password||cloudant password
-riak.host|| riak ip address
-riak.port|| riak port
 jsonstore.rdd.partitions|5|the number of partitions intent used to drive JsonStoreRDD loading query result in parallel. The actual number is calculated based on total rows returned and satisfying maxInPartition and minInPartition
 jsonstore.rdd.maxInPartition|-1|the max rows in a partition. -1 means unlimited
 jsonstore.rdd.minInPartition|10|the min rows in a partition.
@@ -203,7 +185,7 @@ Name | Default | Meaning
 --- |:---:| ---
 database||cloudant database name
 index||cloudant search index w/o the database name.only used for load.
-path||riak: search index name; cloudant: as database name if database does not present
+path||cloudant: as database name if database is not present
 
 
 ## Known limitations and improvement areas
