+++ cloudant-spark-sql/src/main/scala/com/cloudant/spark/common/JsonStoreConfigManager.scala
@@ -16,16 +16,7 @@
 package com.cloudant.spark.common
 
 import org.apache.spark.sql.SQLContext
-import play.api.libs.json.JsValue
-import play.api.libs.json.JsSuccess
-import play.api.libs.json.JsError
-import play.api.libs.json.Json
-import scala.util.control.Breaks._
-import play.api.libs.json.JsUndefined
-import java.net.URLEncoder
 import com.typesafe.config.ConfigFactory
-import scala.concurrent.duration.Duration
-import com.cloudant.spark.riak.RiakConfig
 import com.cloudant.spark.CloudantConfig
 
  object JsonStoreConfigManager
@@ -33,8 +24,6 @@ import com.cloudant.spark.CloudantConfig
   val CLOUDANT_HOST_CONFIG = "cloudant.host"
   val CLOUDANT_USERNAME_CONFIG = "cloudant.username"
   val CLOUDANT_PASSWORD_CONFIG = "cloudant.password"
-  val RIAK_HOST_CONFIG = "riak.host"
-  val RIAK_PORT_CONFIG = "riak.port"
   
   val PARTITION_CONFIG = "jsonstore.rdd.partitions"
   val MAX_IN_PARTITION_CONFIG = "jsonstore.rdd.maxInPartition"
@@ -44,7 +33,6 @@ import com.cloudant.spark.CloudantConfig
   val BULK_SIZE_CONFIG = "jsonstore.rdd.bulkSize"
   
   val configFactory = ConfigFactory.load()
-  import java.util.concurrent.TimeUnit._
 
   val ROOT_CONFIG_NAME = "spark-sql"
   val rootConfig = configFactory.getConfig(ROOT_CONFIG_NAME) 
@@ -55,7 +43,7 @@ import com.cloudant.spark.CloudantConfig
   val defaultConcurrentSave = rootConfig.getInt(CONCURRENT_SAVE_CONFIG)
   val defaultBulkSize = rootConfig.getInt(BULK_SIZE_CONFIG)
   
-  def getConfig(context: SQLContext, dbName: String, indexName:String = null): JsonStoreConfig = {
+  def getConfig(context: SQLContext, dbName: String, indexName:String = null): CloudantConfig = {
       val sparkConf = context.sparkContext.getConf
       implicit val total = sparkConf.getInt(PARTITION_CONFIG,defaultPartitions)
       implicit val max = sparkConf.getInt(MAX_IN_PARTITION_CONFIG,defaultMaxInPartition)
@@ -63,19 +51,13 @@ import com.cloudant.spark.CloudantConfig
       implicit val requestTimeout =sparkConf.getLong(REQUEST_TIMEOUT_CONFIG,defaultRequestTimeout)
       implicit val concurrentSave =sparkConf.getInt(CONCURRENT_SAVE_CONFIG,defaultConcurrentSave)
       implicit val bulkSize =sparkConf.getInt(BULK_SIZE_CONFIG,defaultBulkSize)
-  
+
       if (sparkConf.contains(CLOUDANT_HOST_CONFIG))
       {
         val host = sparkConf.get(CLOUDANT_HOST_CONFIG)
         val user = sparkConf.get(CLOUDANT_USERNAME_CONFIG)
         val passwd = sparkConf.get(CLOUDANT_PASSWORD_CONFIG)
-        return CloudantConfig(host,  dbName, indexName)(user, passwd, total, max, min,requestTimeout,concurrentSave, bulkSize)
-      }
-      if (sparkConf.contains(RIAK_HOST_CONFIG))
-      {
-        val host = sparkConf.get(RIAK_HOST_CONFIG)
-        val port = sparkConf.get(RIAK_PORT_CONFIG)
-        return RiakConfig(host, port, dbName)(partitions=total, maxInPartition=max, minInPartition=min,requestTimeout=requestTimeout,concurrentSave=concurrentSave, bulkSize = bulkSize)
+        return new CloudantConfig(host,  dbName, indexName)(user, passwd, total, max, min,requestTimeout,concurrentSave, bulkSize)
       }
       null
   }
@@ -84,7 +66,7 @@ import com.cloudant.spark.CloudantConfig
    * The configuration sequence: DataFrame option override SparkConf override defaults
    */
   
-  def getConfig(context: SQLContext,  parameters: Map[String, String]): JsonStoreConfig = {
+  def getConfig(context: SQLContext,  parameters: Map[String, String]): CloudantConfig = {
     
       val sparkConf = context.sparkContext.getConf
       
@@ -110,13 +92,7 @@ import com.cloudant.spark.CloudantConfig
         val host = parameters.getOrElse(CLOUDANT_HOST_CONFIG,sparkConf.get(CLOUDANT_HOST_CONFIG))
         val user = parameters.getOrElse(CLOUDANT_USERNAME_CONFIG,sparkConf.get(CLOUDANT_USERNAME_CONFIG))
         val passwd = parameters.getOrElse(CLOUDANT_PASSWORD_CONFIG,sparkConf.get(CLOUDANT_PASSWORD_CONFIG))
-        return CloudantConfig(host,  dbName, indexName)(user, passwd, total, max, min,requestTimeout,concurrentSave,bulkSize)
-      }
-      if (sparkConf.contains(RIAK_HOST_CONFIG)  || parameters.contains(RIAK_HOST_CONFIG))
-      {
-        val host = parameters.getOrElse(RIAK_HOST_CONFIG,sparkConf.get(RIAK_HOST_CONFIG))
-        val port = parameters.getOrElse(RIAK_PORT_CONFIG,sparkConf.get(RIAK_PORT_CONFIG))
-        return RiakConfig(host, port, dbName)(partitions=total, maxInPartition=max, minInPartition=min,requestTimeout=requestTimeout,concurrentSave=concurrentSave,bulkSize = bulkSize)
+        return new CloudantConfig(host,  dbName, indexName)(user, passwd, total, max, min,requestTimeout,concurrentSave,bulkSize)
       }
       null
   }
